Fire events when a player connects or disconnects
We provide connected and disconnected events for the client (when they
have successfully connected to the server) and server (whenever a new
client connects).

There are some caveats with this implementation, namely that the
client-side events are fired on the network thread. It may be worth
investigating if there are better places the hooks could be placed.