Fix BlockRenderLayerMapImpl no longer working in 19w42a. Move some LOG.warns to InvalidParameterException
In 19w42a new class (class_4699) that contains 3 maps for Item, Block and Fluid RenderLayers.
Before these maps were in the RenderLayer class causing them to be populated well before this API had a chance to get to handling the consumers.

I have tweaked the impl to handle RenderLayers being added before initialize is called and after, this removes the need to ensure the class load order is correct.